+++ backend/src/api/mod.rs
@@ -1,4 +1,6 @@
 /// Обработчик API для репозиториев
 pub mod repo;
 /// Обработчик API для пользователей
-pub mod user; 
+pub mod user;
+/// Обработчик API для Git HTTP протокола
+pub mod git;
