+++ backend/src/models/mod.rs
@@ -1,4 +1,4 @@
 pub mod user;
 pub mod repository;
 pub mod db;
-// pub mod notification;
+pub mod notification;
